requirements.txt
certifi==2021.10.8
#charset-normalizer==2.0.12
#idna==3.3
requests==2.27.1
#telegram==0.0.1
urllib3==1.26.9
python-telegram-bot==12.7